pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "clix-cli"
readme = "README.md"
authors = [{name = "Siddharth Bayapureddy" , email = "siddharthbayapureddy@gmail.com"}]
version = "0.1.0"
description = "A terminal-based AI companion."
requires-python = ">=3.10"
dependencies = [
    "typer",
    "litellm",
    "rich",
    "python-dotenv"
]


[project.scripts]
clix = "clix.main:app"